rename serv package to metrics
